Server manager finished. CLI Wrapper fixed.
